First batch for edition of photos

+++ app/src/main/java/com/h/pixeldroid/interfaces/EditImageFragmentListener.kt
@@ -3,9 +3,9 @@ package com.h.pixeldroid.interfaces
 interface EditImageFragmentListener {
     fun onBrightnessChange(brightness: Int)
 
-    fun onSaturationChange(saturation: Int)
+    fun onSaturationChange(saturation: Float)
 
-    fun onContrastChange(contrast: Int)
+    fun onContrastChange(contrast: Float)
 
     fun onEditStarted()
